Security Part IV: Content restriction; Correction in JWT Filter

+++ src/main/java/stock/service/UserService.java
@@ -4,12 +4,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort.Direction;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import stock.domain.User;
+import stock.domain.enumerator.Profile;
+import stock.exception.AuthorizationException;
 import stock.exception.IllegalArgumentException;
 import stock.repository.UserRepository;
+import stock.security.UserDetailsImpl;
 
 @Service
 public class UserService {
@@ -49,6 +53,12 @@ public class UserService {
 	
 	public User update(User user, Long id) {
 		
+		UserDetailsImpl authenticatedUser = authenticated();		
+		
+		if (authenticatedUser == null || (!id.equals(authenticatedUser.getId()) && !authenticatedUser.hasRole(Profile.ROLE_ADMIN))) {
+			throw new AuthorizationException("Access denied");
+		}		
+		
 		if (user.getId() != id) {
 			user.setId(id);
 		}
@@ -60,12 +70,27 @@ public class UserService {
 	
 	public void delete(Long id) {
 		
+		UserDetailsImpl authenticatedUser = authenticated();		
+		
+		if (authenticatedUser == null || authenticatedUser.hasRole(Profile.ROLE_ADMIN) || id.equals(authenticatedUser.getId())) {
+			throw new AuthorizationException("Access denied");
+		}		
+		
 		userRepository.deleteById(id);
 	}
 	
 	private String encode(String password) {
 		
 		return passwordEncoder.encode(password);
+	}
+	
+	public static UserDetailsImpl authenticated() {
+		
+		try {
+			return (UserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		} catch (Exception e) {
+			return null;
+		}
 	}
 	
 }
